Add priority functionality to depots
When creating a request, request and buffer depots now ask depots on the
network for their priority. All types of depot except fuel and request
implement a priority getter, which checks for an attached circuit
writer. If it exists, the signal "0" is read from both wires and merged.
This signal is returned as the priority. If no circuit writer exists, or
if the signal "0" is not present, 0 is returned. The highest priority group
whose contents satisfy the minimum request size is chosen to receive the
request, and the normal heuristic is run against this group to determine
the specific request target.